// src/app/models/article.ts
export interface Article {
    id?: number;
    artCode: string;
    artIsNegoce: boolean;
    artLibC: string;
    artLibL: string;
    artLstSoc: string;
    artPu: number;
    artPua: number;
    artQteMin: number;
    artQteMinA: number;
    artType: string;
    uniteAchat: string;
    datCrea?: string;
    datModi?: string;
    quiCrea?: string;
    quiModi?: string;
    unite: string;
  }
  